Adding Docker to building.rst
This is the first draft, not ready for merge yet but I wanted to get it started.  The theia build fails about 30% through on this one and I'm not yet sure why.   Likely an issue with on of the depenency versions since I got 0.7 and 'latest' to build on Docker and on my laptop.  Chris, I'm hoping this will help with users getting theia up and running a little smoother.  I'll double check the dependency versions and make an issue if i still can't get it to build.